requirements.txt
django==1.8
django-bootstrap3==5.2.0
django-picklefield==0.3.1
names==0.3.0
